refactor: extract types and replay_buffer from core/mod.rs
Split self-contained pieces out of the 1,750-line core/mod.rs:

- core/types.rs (210 lines): StitchCoreError, RenderOutcome,
  ReplayFrame, StitchCoreConfig, StackedReplayRecorder trait,
  StackedReplayGpuRecorder trait
- core/replay_buffer.rs (117 lines): ReplayBuffer bounded-duration
  ring with all its methods

core/mod.rs: 1,750 -> 1,455 lines. StitchCore struct and all its
methods remain in mod.rs. Re-exports preserve external API.

8 tests pass, clippy clean.
